betis-arbitrage/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
cloudscraper>=1.2.71
python-dotenv>=0.20.0